Issue Id   : #229 Message    : Update the Technical Question,Written Test And Total marks in interview Form Transcript : Updated implementation of technical questions total mark in interview form

+++ HumanitarianAssistance.Application/Project/Commands/Create/AddInterviewDetailsCommand.cs
@@ -33,7 +33,7 @@ namespace HumanitarianAssistance.Application.Project.Commands.Create
         public bool ExpectationTransport { get; set; }
         public bool ExpectationMeal { get; set; }
         public double ProfessionalCriteriaMark { get; set; }
-        public int MarksObtain { get; set; }
+        public double MarksObtain { get; set; }
         public double TotalMarksObtain { get; set; }
     }
 }
